protocol: changed operator_stake_default to u128 inline with other stake data types

+++ protocol/protocol/contracts/lib.rs
@@ -168,7 +168,7 @@ pub mod prosopo {
         operators: Mapping<AccountId, Operator>,
         operator_accounts: Lazy<Vec<AccountId>>,
         //disputes: Mapping<u64, Dispute>
-        operator_stake_default: u64,
+        operator_stake_default: u128,
         operator_fee_currency: Hash,
         dapp_users: Mapping<AccountId, User>,
         dapp_user_accounts: Lazy<Vec<AccountId>>,
